requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0
google-genai
tavily-python>=0.3.0
streamlit>=1.35.0
pydantic>=2.7.0
python-dotenv>=1.0.0
typing-extensions>=4.9.0